+++ Raytrace/Dielectric.cs
@@ -4,15 +4,15 @@ namespace RayTrace
 {
     public class Dielectric : Material
     {
-        public float RefractionIndex { get; set; }
+        public double RefractionIndex { get; set; }
 
-        float Schlick(float cosine, float ref_idx)
+        double Schlick(double cosine, double ref_idx)
         {
             var r0 = (1 - ref_idx) / (1 + ref_idx);
             r0 = r0 * r0;
-            return r0 + (1 - r0) * (float)Math.Pow(1 - cosine, 5);
+            return r0 + (1 - r0) * Math.Pow(1 - cosine, 5);
         }
-        public bool Refract(Vector3 v, Vector3 n, float niOverNt, out Vector3 refracted)
+        public bool Refract(Vector3 v, Vector3 n, double niOverNt, out Vector3 refracted)
         {
             var uv = Vector3.UnitVector(v);
             var dt = Vector3.Dot(uv, n);
@@ -20,7 +20,7 @@ namespace RayTrace
 
             if (discriminant > 0)
             {
-                refracted = niOverNt * (v - n * dt) - n * (float)Math.Sqrt(discriminant);
+                refracted = niOverNt * (v - n * dt) - n * Math.Sqrt(discriminant);
                 return true;
             }
             refracted = new Vector3(0, 0, 0);
@@ -31,11 +31,11 @@ namespace RayTrace
         {
             Vector3 outwardNormal;
             var reflected = Reflect(r.Direction, record.normal);
-            float niOverNt;
+            double niOverNt;
             attenuation = new Vector3(1, 1, 1);
             Vector3 refracted;
-            float reflectProb;
-            float cosine;
+            double reflectProb;
+            double cosine;
 
             if (Vector3.Dot(r.Direction, record.normal) > 0)
             {
