+++ Raytrace/FixFocus.cs
@@ -2,14 +2,14 @@
 {
     public class FixFocus : Focus
     {
-        private float _distance;
-        public float Distance
+        private double _distance;
+        public double Distance
         {
             get => _distance;
             set { _distance = value; }
         }
 
-        public override float GetFocusDistance(Camera camera)
+        public override double GetFocusDistance(Camera camera)
         {
             return Distance;
         }
